Give retained frame state and observations an explicit runtime module
